+++ src/utils/ids-device-env-specs-utils/ids-device-env-specs-utils.d.ts
@@ -1,0 +1,14 @@
+interface Specs {
+  currentBrowser: string,
+  browserVersion: string,
+  browserMajorVersion: string,
+  isMobile: boolean,
+  os: string,
+  currentOSVersion: string,
+  idsVersion: string,
+  cookiesEnabled: boolean,
+  platform: string,
+  browserLanguage: string
+}
+
+export function getSpecs(): Specs;